Develop vuong v3

+++ MiniProject/CommodityModule/src/main/java/com/haonguyen/CommodityService/service/CommodityService.java
@@ -1,6 +1,5 @@
 package com.haonguyen.CommodityService.service;
 
-
 import com.haonguyen.CommodityService.dto.*;
 import com.haonguyen.CommodityService.iservice.ICommodityService;
 import com.haonguyen.CommodityService.mapper.ICommodityMapper;
@@ -9,7 +8,9 @@ import com.haonguyen.CommodityService.repository.ICommodityRepository;
 import com.mini_project.CoreModule.entity.CommodityEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestOperations;
 import org.springframework.web.client.RestTemplate;
+
 
 import java.util.List;
 import java.util.UUID;
@@ -20,57 +21,31 @@ public class CommodityService implements ICommodityService {
     private final ICommodityRepository iCommodityRepository;
     private final ICommodityMapper commodityMapper;
 
-    public CommodityService(ICommodityRepository commodityRepository,
-                            ICommodityMapper commodityMapper) {
-        this.iCommodityRepository = commodityRepository;
+    public CommodityService(ICommodityRepository iCommodityRepository, ICommodityMapper commodityMapper) {
+        this.iCommodityRepository = iCommodityRepository;
         this.commodityMapper = commodityMapper;
     }
 
 
-    @Override
-      public CommodityCreateDto addCommodity(CommodityCreateDto commodityCreateDto) {
-        if (commodityCreateDto == null) {
-            return null;
-        }
-        ICommodityMapper iCommodityMapper = new ICommodityMapperImpl();
-        CommodityEntity commodityEntity = iCommodityMapper.fromCreateToEntity(commodityCreateDto);
-        iCommodityRepository.saveAndFlush(commodityEntity);
-        CommodityCreateDto commodityCreateDto1 =iCommodityMapper.toCreateDto(commodityEntity);
-        return commodityCreateDto1;
-    }
-
-    @Override
-    public CommodityUpdateDto updateCommodity(CommodityUpdateDto commodityUpdateDto){
-        if (commodityUpdateDto == null){
-            return null;
-        }
-        ICommodityMapper iCommodityMapper = new ICommodityMapperImpl();
-
-        CommodityEntity commodityEntity = iCommodityMapper.fromUpdateToEntity(commodityUpdateDto);
-
-        iCommodityRepository.save(commodityEntity);
-
-        CommodityUpdateDto commodityUpdateDto1 =iCommodityMapper.toUpdateDto(commodityEntity);
-
-        return commodityUpdateDto1;
-    }
+    /**
+     * xoa hang theo dieu kien,ton kho,khong co tron phieu nhap
+     */
     @Autowired
-    private RestTemplate restTemplate;
+    RestTemplate restTemplate;
     @Override
     public void deleteCommodity(UUID id) {
-        String checkIdCommodityURL = "http://localhost:9002/v1/api/export/check-id-commodity/";
-        boolean ABC = restTemplate.getForObject(checkIdCommodityURL + id,Boolean.class);
-        iCommodityRepository.deleteById(id);
-    }
-//    @Override
-//    public CommodityEntity del(ItemReceiptDTO list) {
-//        String sourceCommodityURL = "http://COMMODITY-SERVICE/commodity/getId/";
-//        CommodityDTO resultCommodityDto = restTemplate.getForObject(sourceCommodityURL + list.getIdCommodity(), CommodityDTO.class);
-//        CommodityDTOMapper commodityDTOMapper = new CommodityDTOMapperImpl();
-//        CommodityEntity commodityEntity = commodityDTOMapper.toCommodityEntity(resultCommodityDto);
-//        return commodityEntity;
-//    }
+        Double commodityInWarehouse = iCommodityRepository.checkCommodityInWarehouseById(id).getInventoryNumber();
+        String checkIdCommodityURL = "http://localhost:8112/v1/api/export/check-id-commodity/";
+        String ABC = restTemplate.getForObject(checkIdCommodityURL + id,String.class);
+        if(ABC.equals("true") && commodityInWarehouse == 0 )
+            iCommodityRepository.deleteById(id);
 
+
+
+    }
+    /**
+     *tim kiem da dieu kien
+     */
     @Override
     public List<CommoditySearchDto> searchCommodity(String key) {
         if (key==null) {
@@ -84,19 +59,62 @@ public class CommodityService implements ICommodityService {
         }
     }
 
+    /**
+     *Them hang
+     */
+    @Override
+      public CommodityCreateDto addCommodity(CommodityCreateDto commodityCreateDto) {
+        if (commodityCreateDto == null) {
+            return null;
+        }
+        ICommodityMapper iCommodityMapper = new ICommodityMapperImpl();
+        CommodityEntity commodityEntity = iCommodityMapper.fromCreateToEntity(commodityCreateDto);
+        iCommodityRepository.saveAndFlush(commodityEntity);
+        CommodityCreateDto commodityCreateDto1 =iCommodityMapper.toCreateDto(commodityEntity);
+        return commodityCreateDto1;
+    }
+
+    /**
+     *Cap nhat hang
+     */
+    @Override
+    public CommodityUpdateDto updateCommodity(CommodityUpdateDto commodityUpdateDto, UUID idCommodity){
+        if (commodityUpdateDto == null){
+            return null;
+        }
+        commodityUpdateDto.setId(idCommodity);
+        ICommodityMapper iCommodityMapper = new ICommodityMapperImpl();
+        CommodityEntity commodityEntity = iCommodityMapper.fromUpdateToEntity(commodityUpdateDto);
+        iCommodityRepository.save(commodityEntity);
+        CommodityUpdateDto commodityUpdateDto1 =iCommodityMapper.toUpdateDto(commodityEntity);
+        return commodityUpdateDto1;
+    }
+
+    /**
+     *Lay thong tin theo loai hang
+     */
     @Override
     public List<TypeOfCommodityDto> findCommodityByIdTypeOfCommodity(UUID idTypeOfCommodity) {
         ICommodityMapper iCommodityMapper = new ICommodityMapperImpl();
         List<TypeOfCommodityDto> typeOfCommodityDto = iCommodityMapper
                 .toFindSectorId(iCommodityRepository.findCommodityByIdTypeOfCommodity(idTypeOfCommodity));
         return typeOfCommodityDto;
-
+    /**
+     * lay thong tin thue
+     */
     }
     @Override
     public TypeAndTaxCommodityAPI getTypeTaxCommodity(UUID idCommodity) {
-        return iCommodityRepository.getTypeTaxCommodity(idCommodity);
+        if (idCommodity!= null) {
+            TypeAndTaxCommodityAPI typeAndTaxCommodityAPI = iCommodityRepository.getTypeTaxCommodity(idCommodity);
+            return typeAndTaxCommodityAPI;
+        }
+        else
+            return null;
     }
-
+    /**
+     *Lay thong tin theo id
+     */
     @Override
     public CommodityCreateDto CommodityById(UUID id) {
         ICommodityMapper iCommodityMapper= new ICommodityMapperImpl();
